# docs/api.yml
openapi: 3.1.1
info:
  title: MeetingService API
  description: Handles API calls for creation of new meetings and the meeting agenda / current status of meeting.
  version: 0.1.0

servers:
  - url: http://meeting-service.ltu-m7011e-2.se

paths:
  /code/{code}:
    get: 
      summary: Get meeting uuid from code
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
            minLength: 6
            maxLength: 6
      responses:
        "200":
          description: Successfully got meeting uuid from code
          content:
            text/plain:
              schema:
                type: string
                format: uuid
        "404":
          description: Meeting code does not match any meetings

  /meetings:
    post:
      summary: Create a new meeting.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeetingRequest"
      responses:
        "201":
          description: Meeting successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}/:  
    get:
      summary: Get meeting info     
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successfully got meeting info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
    patch:
      summary: Update properties of a meeting (e.g., current agenda item)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeetingRequest"
      responses:
        "200":
          description: Meeting successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}/agenda:
    post:
      summary: Add a point to the agenda
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgendaItemRequest"
      responses:
        "201":
          description: Agenda item added successfully
    get:
      summary: Get all agenda points
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Agenda item added successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgendaItem"
components:
  schemas:
    Meeting:
      type: object
      description: A meeting.
      properties:
        meeting_id:
          type: string
          format: uuid
        meeting_code:
          type: string
          minLength: 6
          maxLength: 6
          example: "123456"
        meeting_name:
          type: string
          example: "The Best Meeting!"
        current_item:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/AgendaItem"
        
      required:
        - meeting_id
        - meeting_name

    CreateMeetingRequest:
      type: object
      properties:
        meeting_name:
          type: string
        
      required:
        - meeting_name
        - position_name

    UpdateMeetingRequest:
      type: object
      description: Fields that may be updated in a meeting
      properties:
        current_item:
          type: integer
          description: Index of the currently active agenda item
      additionalProperties: false

    AgendaItem:
      oneOf:
        - $ref: "#/components/schemas/ElectionAgendaItem"
        - $ref: "#/components/schemas/MotionAgendaItem"
        - $ref: "#/components/schemas/InfoAgendaItem"
      discriminator:
        propertyName: type
        mapping:
          election: "#/components/schemas/ElectionAgendaItem"
          motion: "#/components/schemas/MotionAgendaItem"
          info: "#/components/schemas/InfoAgendaItem"

    ElectionAgendaItem:
      type: object
      description: A election item in a meeting's agenda.
      properties:
        type:
          type: string
          const: "election"
        title:
          type: string
        positions:
          type: array
          items: 
            type: string

    MotionAgendaItem:
      type: object
      description: A motion item in a meeting's agenda.
      properties:
        type:
          type: string
          const: "motion"
        title:
          type: string
        description:
          type: string
        baseMotions:
          type: array
          items: 
            $ref: "#/components/schemas/BaseMotion"

    BaseMotion:
      type: object
      description: A motion included in the agenda
      properties:
        owner:
          type: string
        motion:
          type: string
        
    InfoAgendaItem:
      type: object
      description: A generic info item in a meeting's agenda.
      properties:
        type:
          type: string
          const: "info"
        title: 
          type: string
        description:
          type: string

    CreateAgendaItemRequest:
      type: object
      properties:
        item:
          $ref: "#/components/schemas/AgendaItem"
      required:
        - item